[ALPS09686298] Coverity fix for gpu dvfs kernel 6.12 code
[Description]
Coverity fix for gpu dvfs kernel 6.12 code:
(i)  Negative loop bound -> Type change for the loop idx
(ii) Unitentional integer overflow -> type cast unsigned long long
(iii)Overflowed constant -> Add if condition to exclude the case

[Test]
Local build pass & Boot up pass

MTK-Commit-Id: 0f747229113bb6c94bab6bc159b084c97ad726e8

CR-Id: ALPS09686298
Feature: GPU DVFS Policy

+++ drivers/gpu/mediatek/ged/src/ged_kpi.c
@@ -1518,7 +1518,7 @@ static void ged_kpi_update_soc_timer(struct GED_KPI_HEAD *psHead, unsigned int u
 			/* compare overdue timer to check uncomplete soc timer*/
 			info_soc_overdue +=
 					(g_ged_eb_uncomplete_info.last_uncomplete_target / 1000) * soc_timer_unit;
-			psHead_soc_overdue += head_t_gpu_target * soc_timer_unit;
+			psHead_soc_overdue += ((unsigned long long)head_t_gpu_target * soc_timer_unit);
 			if (psHead_soc_overdue < info_soc_overdue)
 				timer_change = 1;
 		}
